refactor : bot layer
- refactor discord bot related feature to bot/usecase layer
