--- bot.py
@@ -1,21 +1,0 @@
-import discord
-from discord.ext import commands
-from view import ServerControlView
-
-
-class MinecraftServerBot(commands.Bot):
-    def __init__(self):
-        intents = discord.Intents.default()
-        intents.message_content = True
-
-        super().__init__(command_prefix=commands.when_mentioned_or('!'), intents=intents)
-
-    async def setup_hook(self) -> None:
-        self.add_view(ServerControlView())
-
-    async def on_ready(self):
-        print(f'Logged in as {self.user} (ID: {self.user.id})')
-        print('------')
-
-
-bot = MinecraftServerBot()
