+++ src/bot/__init__.py
@@ -1,0 +1,2 @@
+from .minecraft_server_bot import MinecraftServerBot
+from .cogs.minecaft_cog import Minecraft
